Removes _x64 from directory.
