+++ Shivers Randomizer/App.xaml
@@ -1,0 +1,5 @@
+﻿<Application x:Class="Shivers_Randomizer.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    <Application.Resources />
+</Application>
