+++ Shivers Randomizer/AttachPopup.xaml
