+++ Shivers Randomizer/AttachPopup.xaml.cs
