+++ Shivers Randomizer/MainWindow.xaml.cs
