+++ Shivers Randomizer/Message.xaml
