+++ Shivers Randomizer/Message.xaml.cs
