+++ Shivers Randomizer/PotLocation.cs
