+++ Shivers Randomizer/Properties/launchSettings.json
