+++ Shivers Randomizer/resources/DefaultMap.json
