+++ Shivers Randomizer/utils/AppHelpers.cs
