+++ Shivers Randomizer/utils/DebugSettings.cs
